fix: update exception handling, boost ES availability, and fix collection item sync

+++ sec-server/src/main/java/com/sec/domain/es/ItemDocument.java
@@ -33,8 +33,8 @@ public class ItemDocument {
     @Field(type = FieldType.Double)
     private BigDecimal originalPrice;
 
-    @Field(type = FieldType.Integer)
-    private Integer categoryId;
+    @Field(type = FieldType.Long)
+    private Long categoryId;
 
     private String cover;
